Break apart Event definitions from EventT definitions
