+++ warp-test.cabal
@@ -18,6 +18,7 @@ library
   ghc-options:         -Wall
   exposed-modules:     Css
                      , Events
+                     , EventT
                      , Lib
                      , Mailer
                      , Middleware
